+++ src/datasource/CustomerDatasource.ts
@@ -1,5 +1,5 @@
 import { AxiosResponse } from "axios";
-import { BASE_URL, httpClient } from "../config/develop-config";
+import { BASE_URL, httpClient, NAV_URL } from "../config/develop-config";
 
 const baseUrl = `${BASE_URL}/auth`;
 
@@ -24,4 +24,17 @@ const getZones = async (company: string) => {
     .catch((err) => console.log(err));
 };
 
-export { getCustomers, getCustomersById, getZones };
+const checkPhoneByShop = async (params: object) => {
+  return await httpClient
+    .post(`${NAV_URL}/customer/check-customer`, params)
+    .then((res: AxiosResponse) => res.data)
+    .catch((err) => console.log(err));
+};
+const checkPhoneAllShop = async (params: object) => {
+  return await httpClient
+    .post(`${NAV_URL}/customer/check-all-customer`, params)
+    .then((res: AxiosResponse) => res.data)
+    .catch((err) => console.log(err));
+};
+
+export { getCustomers, getCustomersById, getZones, checkPhoneByShop, checkPhoneAllShop };